requirements.txt
rpaframework==11.6.3
